requirements.txt
# requirements.txt

streamlit==1.28.0
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.4
python-dotenv==1.0.0
google-search-results==2.4.2
Pillow==10.0.0
googletrans==4.0.0-rc1
wordcloud
matplotlib

# sqlite3 is part of Python standard library (no need to install)
